+++ packages/source/src/parser/raster-tile.ts
@@ -1,5 +1,9 @@
-import { TileLoadParams, TilesetManagerOptions } from '@antv/l7-core';
-import { getImage, getURLFromTemplate } from '@antv/l7-utils';
+import {
+  getImage,
+  getURLFromTemplate,
+  TileLoadParams,
+  TilesetManagerOptions,
+} from '@antv/l7-utils';
 import { IParserData, IRasterTileParserCFG } from '../interface';
 
 const DEFAULT_CONFIG: Partial<TilesetManagerOptions> = {
