+++ packages/utils/src/tileset-manager/index.ts
@@ -1,3 +1,4 @@
 export * from './utils/tile-url';
+export * from './types';
 export { Tile } from './tile';
 export { TilesetManager } from './tileset-manager';
