+++ packages/utils/src/tileset-manager/tile.ts
@@ -1,26 +1,11 @@
 import bboxPolygon from '@turf/bbox-polygon';
-import { Bounds } from './types';
+import {
+  LoadTileDataStatus,
+  TileBounds,
+  TileLoadDataOptions,
+  TileOptions,
+} from './types';
 import { getTileWarpXY, tileToBounds } from './utils/lonlat-tile';
-
-type TileOptions = { x: number; y: number; z: number; tileSize: number };
-
-export type TileLoadParams = TileOptions & {
-  bounds: Bounds;
-  signal: AbortSignal;
-};
-
-type TileLoadDataOptions = {
-  getData: (tile: TileLoadParams) => Promise<any>;
-  onLoad: (tile: Tile) => void;
-  onError: (error: Error, tile: Tile) => void;
-};
-
-enum LoadTileDataStatus {
-  Loading = 'Loading',
-  Loaded = 'Loaded',
-  Failure = 'Failure',
-  Cancelled = 'Cancelled',
-}
 
 /**
  * 单个瓦片
@@ -98,7 +83,7 @@ export class Tile {
     const [minLng, minLat, maxLng, maxLat] = this.bounds;
     const center = [(maxLng - minLng) / 2, (maxLat - minLat) / 2] as const;
 
-    const polygon = bboxPolygon(this.bounds as Bounds, {
+    const polygon = bboxPolygon(this.bounds as TileBounds, {
       properties: {
         key: this.key,
         bbox: this.bounds,
