+++ packages/utils/src/tileset-manager/tileset-manager.ts
@@ -5,19 +5,9 @@ import {
   NOOP,
   UPDATE_TILE_STRATEGIES,
 } from './const';
-import { Tile, TileLoadParams } from './tile';
-import { Bounds, UpdateTileStrategy } from './types';
+import { Tile } from './tile';
+import { TilesetManagerOptions, UpdateTileStrategy } from './types';
 import { getTileIndices } from './utils/lonlat-tile';
-
-export type TilesetManagerOptions = {
-  tileSize: number;
-  zoomOffset: number;
-  minZoom: number;
-  maxZoom: number;
-  extent: Bounds;
-  getTileData: (tile: TileLoadParams) => any;
-  updateStrategy: UpdateTileStrategy | ((tiles: Tile[]) => void);
-};
 
 /**
  * 管理瓦片数据
